Moved polygon series into subfolder.
