+++ js/masters/highcharts-more.src.js
@@ -18,7 +18,7 @@ import '../Series/GaugeSeries.js';
 import '../Series/BoxPlot/BoxPlotSeries.js';
 import '../Series/ErrorBar/ErrorBarSeries.js';
 import '../Series/WaterfallSeries.js';
-import '../Series/PolygonSeries.js';
+import '../Series/Polygon/PolygonSeries.js';
 import '../Series/Bubble/BubbleSeries.js';
 import '../Series/PackedBubble/PackedBubbleSeries.js';
 import '../Extensions/Polar.js';
